create subpackages out of the status package
