+++ pkg/status/clusteragent/status_apiserver.go
@@ -1,11 +1,11 @@
 // Unless explicitly stated otherwise all files in this repository are licensed
 // under the Apache License Version 2.0.
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2016-present Datadog, Inc.
+// Copyright 2023-present Datadog, Inc.
 
 //go:build kubeapiserver
 
-package status
+package clusteragent
 
 import (
 	"fmt"
@@ -32,7 +32,8 @@ func getLeaderElectionDetails() map[string]string {
 	return leaderElectionStats
 }
 
-func getDCAStatus() map[string]string {
+// GetDCAStatus collect the DCA agent information and return it in a map
+func GetDCAStatus() map[string]string {
 	clusterAgentDetails := make(map[string]string)
 
 	dcaCl, err := clusteragent.GetClusterAgentClient()
