+++ pkg/status/clusteragent/status_no_apiserver.go
@@ -1,11 +1,11 @@
 // Unless explicitly stated otherwise all files in this repository are licensed
 // under the Apache License Version 2.0.
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2016-present Datadog, Inc.
+// Copyright 2023-present Datadog, Inc.
 
 //go:build !kubeapiserver
 
-package status
+package clusteragent
 
 import (
 	"github.com/DataDog/datadog-agent/pkg/util/log"
@@ -16,7 +16,8 @@ func getLeaderElectionDetails() map[string]string {
 	return nil
 }
 
-func getDCAStatus() map[string]string {
+// GetDCAStatus empty function for agents not running in a  k8s environment
+func GetDCAStatus() map[string]string {
 	log.Info("Not implemented")
 	return nil
 }
