+++ pkg/status/hostname/hostname.go
@@ -1,0 +1,20 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2023-present Datadog, Inc.
+
+// Package hostname fetch information needed to render the 'hostname' section of the status page.
+package hostname
+
+import (
+	"encoding/json"
+	"expvar"
+)
+
+// PopulateStatus populates the status stats
+func PopulateStatus(stats map[string]interface{}) {
+	hostnameStatsJSON := []byte(expvar.Get("hostname").String())
+	hostnameStats := make(map[string]interface{})
+	json.Unmarshal(hostnameStatsJSON, &hostnameStats) //nolint:errcheck
+	stats["hostnameStats"] = hostnameStats
+}
