+++ pkg/status/systemprobe/systemprobe_unsupported.go
@@ -5,10 +5,11 @@
 
 //go:build !process
 
-package status
+// Package systemprobe fetch information about the system probe
+package systemprobe
 
-// GetSystemProbeStats returns a notice that it is not supported on systems that do not at least build the process agent
-func GetSystemProbeStats(_ string) map[string]interface{} {
+// GetStatus returns a notice that it is not supported on systems that do not at least build the process agent
+func GetStatus(_ string) map[string]interface{} {
 	return map[string]interface{}{
 		"Errors": "System Probe is not supported on this system",
 	}
